% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions (2).R, R/helper_functions.R
\name{get_dois_from_paper}
\alias{get_dois_from_paper}
\title{get_dois_from_paper}
\usage{
get_dois_from_paper(HTML, DOI = "")

get_dois_from_paper(HTML, DOI = "")
}
\arguments{
\item{HTML}{Use the website of the paper}

\item{DOI}{Use the DOI number of the paper}
}
\value{



}
\description{
Get all references with a DOI or a PubmedID from a paper

Get all references with a DOI or a PubmedID from a paper
}
\examples{

DF_new = get_dois_from_paper(DOI = "10.1001/jamainternmed.2021.0269")

DF_new = get_dois_from_paper(DOI = "10.1001/jamainternmed.2021.0269")
}
