% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{download_papers}
\alias{download_papers}
\title{download_DOIs}
\usage{
download_papers(
  DOIs,
  wait_pubmed = 2,
  wait_scihub = 10,
  download_folder = "downloads"
)
}
\arguments{
\item{DOIs}{string of DOIs}

\item{wait_pubmed}{in seconds}

\item{wait_scihub}{in seconds}

\item{download_folder}{specify the folder where downloaded papers should go}
}
\value{

}
\description{
Download papers using DOI
}
\examples{

download_papers(DOIs = c("10.1056/NEJMp1608282", "10.1136/bmjopen-2015-008155"), wait_scihub = 5)
}
